requirements.txt
Django>=4.2
djangorestframework
openai
langchain
faiss-cpu
